Tracking
I have added files for tracking, most of this is recycled code. 
This in particular does not cover the velocity of the objects.
I can also not comment on the computation timings, because it does not reuse a lot of code from surf.
This manipulates frames using image segmentation and blob analysis.