Add files via upload
blobMain - attain frames from camera, process frames, pass processed frames to "getDominoInImage" and plot the returned image.

getDominoInImage - perform blob analysis on processed frames, using a threshold area for the blobs to eliminate noise. return an image containing the centroid location and bounding boxes for each domino.

NOTE: "regionprops" in MATLAB may potentially be used to determine the pose of the blobs/the angle between the major axis of the blob and the major axis in the image frame. Look further into "regionprops" to explore this possibility.